Future games will require custom layouts, so will be placing each layout into its own *Layout file. Board Composable will be used to select correct Layout depending on selectedGame.
