+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/game/EasthavenTest.kt
@@ -7,10 +7,9 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.performClick
 import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.ui.MainActivity
-import com.heyzeusv.solitaire.util.Games
 import com.heyzeusv.solitaire.util.TestCards
 import com.heyzeusv.solitaire.util.onNodeWithTextId
-import com.heyzeusv.solitaire.util.switchGame
+//import com.heyzeusv.solitaire.util.switchGame
 import com.heyzeusv.solitaire.util.waitUntilPileCardExists
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
@@ -34,7 +33,7 @@ class EasthavenTest {
 
     @Before
     fun setUp() {
-        composeRule.switchGame(Games.EASTHAVEN)
+//        composeRule.switchGame(Games.EASTHAVEN)
     }
 
     @Test
