+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/game/RussianTest.kt
@@ -4,12 +4,11 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.ui.MainActivity
-import com.heyzeusv.solitaire.util.Games
 import com.heyzeusv.solitaire.util.TestCards
 import com.heyzeusv.solitaire.util.clickOnPileTT
 import com.heyzeusv.solitaire.util.clickOnTableauCard
 import com.heyzeusv.solitaire.util.onNodeWithTextId
-import com.heyzeusv.solitaire.util.switchGame
+//import com.heyzeusv.solitaire.util.switchGame
 import com.heyzeusv.solitaire.util.waitUntilPileCardExists
 import com.heyzeusv.solitaire.util.waitUntilTableauExists
 import dagger.hilt.android.testing.HiltAndroidRule
@@ -34,7 +33,7 @@ class RussianTest {
 
     @Before
     fun setUp() {
-        composeRule.switchGame(Games.RUSSIAN)
+//        composeRule.switchGame(Games.RUSSIAN)
     }
 
     /**
